Clean code Sign In process
+ Move 'sign in' gql mutation to it's single file
+ Create useSignInMutation personal hook
+ Move form onSubmit logic to view component
