+++ src/resolvers/resolvers.js
@@ -1,0 +1,3 @@
+import gql from 'graphql-tag';
+export const typeDefs = {};
+export const resolvers = {};
